[WebView] Continue playback when exiting fullscreen video.
When exiting fullscreen video playback should continue in
embedded mode. This patch fixes a regression introduced in
M40 by reverting to the M39 state:
https://codereview.chromium.org/618013003/patch/350001/360003

The fix landed in master does not longer applied because
master and M40 have diverged substantially. In master,
this issue has been fixed by moving the responsibility of
destroying the ContentVideoView to the content/ layer here:
https://codereview.chromium.org/845193005/

BUG:19346330

TESTED: I have manually verified this change by building
the WebView and checking that the issue does not longer repro.

Change-Id: I7297c626c355e13b559d15c403db60cca817556f