Cherrypick: Don't send extra text changed AX events on Android
This is a clean Cherrypick of 2 CLs from chromium:
https://codereview.chromium.org/920623002
And https://codereview.chromium.org/922883002/

Original description:
Don't send extra text changed AX events on Android

This fix is small and safe, for merging. The reason this
was happening has to do with contenteditable support
and I'll follow up with a cleaner fix once this is merged.

BUG:19294187
Change-Id: Ifd0de22fdad20a917ca74174c90fce8bec7d87ad