Disable orientation locking support in WebView.
The orientation lock code was accidentally inherited from Chrome and
doesn't actually work in WebView. Make it actually fail when JS calls
it, instead of claiming to work but doing nothing, by disabling the
screen orientation provider. This puts WebView in the same state as
desktop chrome which also doesn't set the provider. Tested that this
results in correctly reported errors when JS tries to use the API.

Bug: 19672531
Change-Id: I66a142fc50fa12db418b9465c6f29914633bd9b2